requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
Pillow==10.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7 